requirements.txt
openai>=1.0.0
python-dotenv
fuzzywuzzy
python-Levenshtein
requests
pyttsx3
speechrecognition
selenium
webdriver-manager
